+++ packages/ui/CHANGELOG.md
@@ -6,8 +6,7 @@ The following is a curated list of changes in the Enact ui module, newest change
 
 ### Added
 
-- `ui/Skinnable` to provide themes with a way to apply a base theme styling and skins on top of that
-## Unreleased
+- `ui/Transition` prop `onShow` that fires when transitioning into view a component.
 
 ### Changed